python_scripts/requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.2
